move files to plugin folder
