+++ src/index.js
@@ -24,6 +24,7 @@ export { EnvironmentControls } from './three/controls/EnvironmentControls.js';
 
 // three.js plugins
 export { CesiumIonAuthPlugin } from './three/plugins/CesiumIonAuthPlugin.js';
+export { ImplicitTilingPlugin } from './three/plugins/ImplicitTilingPlugin.js';
 export { GoogleCloudAuthPlugin } from './three/plugins/GoogleCloudAuthPlugin.js';
 export * from './three/plugins/DebugTilesPlugin.js';
