+++ src/three/plugins/ImplicitTilingPlugin.d.ts
@@ -1,5 +1,3 @@
 export class ImplicitTilingPlugin {
-
-	constructor(  );
-
+	constructor();
 }
