+++ src/three/plugins/ImplicitTilingPlugin.js
@@ -1,4 +1,4 @@
-import {SUBTREELoader} from "../loaders/SUBTREELoader.js";
+import {SUBTREELoader} from "../plugins/SUBTREELoader.js";
 
 export class ImplicitTilingPlugin {
 
@@ -16,11 +16,11 @@ export class ImplicitTilingPlugin {
 		if ( tile.implicitTiling ) {	//only for root
 
 			// Store the infos from the tileset
-			tile.availableLevels = tile.implicitTiling.availableLevels;
-			tile.subdivisionScheme = tile.implicitTiling.subdivisionScheme;
+			tile.__availableLevels = tile.implicitTiling.availableLevels;
+			tile.__subdivisionScheme = tile.implicitTiling.subdivisionScheme;
 			tile.__subtreeLevels = tile.implicitTiling.subtreeLevels;
-			tile.subtrees = tile.implicitTiling.subtrees;
-			tile.__subtreeDivider = tile.subdivisionScheme === "QUADTREE" ? 4 : 8;
+			tile.__subtrees = tile.implicitTiling.subtrees;
+			tile.__subtreeDivider = tile.__subdivisionScheme === "QUADTREE" ? 4 : 8;
 			tile.__subtreeUri = tile.implicitTiling.subtrees.uri;
 			tile.__contentUri = uri ?? tile.content?.uri;
 
@@ -45,12 +45,11 @@ export class ImplicitTilingPlugin {
 
 	parseTile(buffer, parseTile, extension  ) {
 		if ( /subtree$/i.test( parseTile.content.uri ) ) {
-			const loader = new SUBTREELoader(null, parseTile, this.tiles.root);
+			const loader = new SUBTREELoader(parseTile, this.tiles.root);
 			loader.parse(buffer);
 			return Promise.resolve()
 		}
 
 	}
 
-
-	}
+}
